backend/requirements.txt
asgiref==3.8.1
djangorestframework==3.15.2
mssql-django==1.5
pyodbc==5.2.0
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.1
Django== 4.2.19
django-mssql-backend==2.8.1
djangorestframework==3.15.2
mssql==1.0.1
django-cors-headers==4.7.0
django-filter==24.3